Remove lock, retrieve max iteration at the end of calculations, make max iterations an ulong.

+++ Calculation.cs
@@ -4,7 +4,7 @@ namespace FractalGenerator
 {
     public struct Calculation
     {
-        public double Iterations;
+        public ulong Iterations;
         public Complex LastZ;
 
         // Used only for Newton
